feat: move constants.ts to constant.service.ts (#482)
