+++ server/src/http/routes/auth/login.controller.ts
@@ -5,7 +5,7 @@ import passport from "passport"
 import { ExtractJwt, Strategy } from "passport-jwt"
 import { Strategy as LocalStrategy } from "passport-local"
 import { mailTemplate } from "../../../assets/index.js"
-import { CFA, ENTREPRISE, etat_utilisateur } from "../../../common/constants.js"
+import { CFA, ENTREPRISE, ETAT_UTILISATEUR } from "../../../services/constant.service.js"
 import { UserRecruteur } from "../../../common/model/index.js"
 import { createMagicLinkToken, createUserRecruteurToken, createUserToken } from "../../../common/utils/jwtUtils.js"
 import config from "../../../config.js"
@@ -136,11 +136,11 @@ export default ({ etablissementsRecruteur, mailer }) => {
 
       switch (user.type) {
         case CFA:
-          if (lastValidationEntry.status === etat_utilisateur.ATTENTE) {
+          if (lastValidationEntry.status === ETAT_UTILISATEUR.ATTENTE) {
             return res.status(400).json({ error: true, reason: "VALIDATION" })
           }
 
-          if (lastValidationEntry.status === etat_utilisateur.DESACTIVE) {
+          if (lastValidationEntry.status === ETAT_UTILISATEUR.DESACTIVE) {
             return res.status(400).json({
               error: true,
               reason: "DISABLED",
@@ -148,11 +148,11 @@ export default ({ etablissementsRecruteur, mailer }) => {
           }
           break
         case ENTREPRISE:
-          if (lastValidationEntry.status === etat_utilisateur.ATTENTE) {
+          if (lastValidationEntry.status === ETAT_UTILISATEUR.ATTENTE) {
             return res.status(400).json({ error: true, reason: "VALIDATION" })
           }
 
-          if (lastValidationEntry.status === etat_utilisateur.DESACTIVE) {
+          if (lastValidationEntry.status === ETAT_UTILISATEUR.DESACTIVE) {
             return res.status(400).json({
               error: true,
               reason: "DISABLED",
